# Cognitive Execution Layer (CEL)/docs/api/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cognitive Execution Layer (CEL) API",
    "description": "Autonomous Engineering API for AI-powered code development and system management",
    "version": "1.0.0",
    "contact": {
      "name": "CEL Development Team",
      "email": "support@cel-project.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.cel-project.org",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/v1/solve-design-constraints": {
      "post": {
        "summary": "Solve design constraints",
        "description": "Find solutions to system design constraints using constraint solving algorithms",
        "tags": ["Constraints"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConstraintSolvingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Constraint solving completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstraintSolvingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/register-agent": {
      "post": {
        "summary": "Register an agent",
        "description": "Register a new autonomous agent in the system",
        "tags": ["Agents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAgentResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConstraintSolvingRequest": {
        "type": "object",
        "required": ["constraints"],
        "properties": {
          "constraints": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "performance"
                },
                "variables": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "predicate": {
                  "type": "string",
                  "example": "responseTime < 100"
                }
              }
            }
          },
          "optimization": {
            "type": "string",
            "example": "minimize_cost"
          }
        }
      },
      "ConstraintSolvingResponse": {
        "type": "object",
        "properties": {
          "solved": {
            "type": "boolean",
            "example": true
          },
          "solution": {
            "type": "object",
            "additionalProperties": true
          },
          "score": {
            "type": "number",
            "example": 0.92
          },
          "iterations": {
            "type": "number",
            "example": 150
          }
        }
      },
      "RegisterAgentRequest": {
        "type": "object",
        "required": ["agentId", "capabilities"],
        "properties": {
          "agentId": {
            "type": "string",
            "example": "code-reviewer-1"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["code_review", "static_analysis"]
          },
          "configuration": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "RegisterAgentResponse": {
        "type": "object",
        "properties": {
          "agentId": {
            "type": "string",
            "example": "code-reviewer-1"
          },
          "registered": {
            "type": "boolean",
            "example": true
          },
          "registrationTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}