# Cognitive Execution Layer (CEL)/src/server/routes/reliability.js
/**
 * Reliability Routes - Handles system reliability, stability, and evolution
 * @module src/server/routes/reliability
 */

/**
 * Create reliability routes
 * @param {Object} app - Express app instance
 * @param {Object} deps - Dependencies
 * @param {Object} deps.stabilityEngine - Stability theory engine instance
 * @param {Object} deps.evolutionEngine - Evolution engine instance
 * @param {Object} deps.antiStagnationEngine - Anti-stagnation engine instance
 * @param {Object} deps.constraintSolver - Constraint solver instance
 */
export function reliabilityRoutes(app, deps) {
  const { stabilityEngine, evolutionEngine, antiStagnationEngine, constraintSolver } = deps;

  /**
   * GET /v1/reliability-info - Get reliability information
   */
  app.get('/v1/reliability-info', async (req, res) => {
    try {
      const stabilityStatus = stabilityEngine ? stabilityEngine.getSystemStabilityStatus() : null;
      const evolutionStatus = evolutionEngine ? evolutionEngine.analyzeTrends() : null;
      const diversityStatus = antiStagnationEngine ? antiStagnationEngine.assessDiversity() : null;

      res.json({
        reliability: {
          stability: stabilityStatus,
          evolution: evolutionStatus,
          diversity: diversityStatus,
          timestamp: new Date().toISOString(),
        },
      });
    } catch (error) {
      console.error('❌ Error in /v1/reliability-info:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get reliability information',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/stability-status - Get stability status
   */
  app.get('/v1/stability-status', async (req, res) => {
    try {
      if (!stabilityEngine) {
        return res.status(503).json({
          error: {
            message: 'Stability engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      // Update metrics with simulated data
      stabilityEngine.updateMetrics({
        conflict: Math.random() > 0.95,
        rollback: Math.random() > 0.97,
        resourceSpike: Math.random() > 0.98,
        architectureDrift: Math.random() > 0.96,
      });

      const stabilityStatus = stabilityEngine.getSystemStabilityStatus();
      const protectionResult = stabilityEngine.activateProtectionMeasures();

      res.json({
        stability: stabilityStatus,
        protection: protectionResult,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/stability-status:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get stability status',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/system-stability-index - Get system stability index
   */
  app.get('/v1/system-stability-index', async (req, res) => {
    try {
      if (!stabilityEngine) {
        return res.status(503).json({
          error: {
            message: 'Stability engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const stabilityIndex = stabilityEngine.calculateStabilityIndex();

      res.json({
        stabilityIndex: stabilityIndex,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/system-stability-index:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get system stability index',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/architecture-evolution-prediction - Get architecture evolution prediction
   */
  app.get('/v1/architecture-evolution-prediction', async (req, res) => {
    try {
      if (!evolutionEngine) {
        return res.status(503).json({
          error: {
            message: 'Evolution engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const forecast = evolutionEngine.forecastArchitectureDecay();
      const trends = evolutionEngine.analyzeTrends();
      const statistics = evolutionEngine.getChangeStatistics();

      res.json({
        forecast,
        trends,
        statistics,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/architecture-evolution-prediction:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get architecture evolution prediction',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/architecture-degradation-check - Check architecture degradation
   */
  app.get('/v1/architecture-degradation-check', async (req, res) => {
    try {
      if (!evolutionEngine) {
        return res.status(503).json({
          error: {
            message: 'Evolution engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const degradationReport = evolutionEngine.detectArchitectureDegradation();

      res.json({
        degradation: degradationReport,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/architecture-degradation-check:', error);
      res.status(500).json({
        error: {
          message: 'Failed to check architecture degradation',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/architecture-entropy - Get architecture entropy
   */
  app.get('/v1/architecture-entropy', async (req, res) => {
    try {
      if (!evolutionEngine) {
        return res.status(503).json({
          error: {
            message: 'Evolution engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const entropy = evolutionEngine.architectureEntropy;
      const history = evolutionEngine.entropyHistory.slice(-20);

      res.json({
        entropy: entropy,
        history: history,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/architecture-entropy:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get architecture entropy',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/anti-stagnation-recommendations - Get anti-stagnation recommendations
   */
  app.get('/v1/anti-stagnation-recommendations', async (req, res) => {
    try {
      if (!antiStagnationEngine) {
        return res.status(503).json({
          error: {
            message: 'Anti-stagnation engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const recommendations = antiStagnationEngine.getImprovementRecommendations();
      const stats = antiStagnationEngine.getUsageStatistics();

      res.json({
        recommendations,
        statistics: stats,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/anti-stagnation-recommendations:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get anti-stagnation recommendations',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/diversity-assessment - Get diversity assessment
   */
  app.get('/v1/diversity-assessment', async (req, res) => {
    try {
      if (!antiStagnationEngine) {
        return res.status(503).json({
          error: {
            message: 'Anti-stagnation engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const diversity = antiStagnationEngine.assessDiversity();

      res.json({
        diversity: diversity,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/diversity-assessment:', error);
      res.status(500).json({
        error: {
          message: 'Failed to assess diversity',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * POST /v1/inject-random-strategy - Inject random strategy
   */
  app.post('/v1/inject-random-strategy', async (req, res) => {
    try {
      if (!antiStagnationEngine) {
        return res.status(503).json({
          error: {
            message: 'Anti-stagnation engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const policy = antiStagnationEngine.injectRandomStrategy();

      res.json({
        policy: policy,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/inject-random-strategy:', error);
      res.status(500).json({
        error: {
          message: 'Failed to inject random strategy',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * POST /v1/inject-random-model - Inject random model
   */
  app.post('/v1/inject-random-model', async (req, res) => {
    try {
      if (!antiStagnationEngine) {
        return res.status(503).json({
          error: {
            message: 'Anti-stagnation engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const policy = antiStagnationEngine.injectRandomModel();

      res.json({
        policy: policy,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/inject-random-model:', error);
      res.status(500).json({
        error: {
          message: 'Failed to inject random model',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * POST /v1/exploration-policy - Get exploration policy
   */
  app.post('/v1/exploration-policy', async (req, res) => {
    try {
      if (!antiStagnationEngine) {
        return res.status(503).json({
          error: {
            message: 'Anti-stagnation engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const context = req.body.context || {};
      const policy = antiStagnationEngine.getExplorationPolicy(context);

      res.json({
        policy: policy,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/exploration-policy:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get exploration policy',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * POST /v1/solve-design-constraints - Solve design constraints
   */
  app.post('/v1/solve-design-constraints', async (req, res) => {
    try {
      const { problem, constraints } = req.body;

      if (!problem) {
        return res.status(400).json({
          error: {
            message: 'Problem definition is required',
            type: 'invalid_request_error',
          },
        });
      }

      if (!constraintSolver) {
        return res.status(503).json({
          error: {
            message: 'Constraint solver not initialized',
            type: 'service_unavailable',
          },
        });
      }

      // Add constraints if provided
      if (constraints && Array.isArray(constraints)) {
        for (const constraint of constraints) {
          constraintSolver.addConstraint(constraint);
        }
      }

      const result = constraintSolver.solveWithConstraints(problem);

      res.json({
        solution: result,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/solve-design-constraints:', error);
      res.status(500).json({
        error: {
          message: 'Failed to solve design constraints',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * GET /v1/constraint-statistics - Get constraint solver statistics
   */
  app.get('/v1/constraint-statistics', async (req, res) => {
    try {
      if (!constraintSolver) {
        return res.status(503).json({
          error: {
            message: 'Constraint solver not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const stats = constraintSolver.getConstraintStatistics();

      res.json({
        statistics: stats,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/constraint-statistics:', error);
      res.status(500).json({
        error: {
          message: 'Failed to get constraint statistics',
          type: 'invalid_request_error',
        },
      });
    }
  });

  /**
   * Agent management routes
   */
  app.post('/v1/register-agent', async (req, res) => {
    try {
      const { agentId, agentSpec } = req.body;

      if (!agentId) {
        return res.status(400).json({
          error: {
            message: 'Agent ID is required',
            type: 'invalid_request_error',
          },
        });
      }

      if (stabilityEngine) {
        stabilityEngine.registerAgent(agentId, agentSpec);
      }

      res.json({
        success: true,
        agentId: agentId,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/register-agent:', error);
      res.status(500).json({
        error: {
          message: 'Failed to register agent',
          type: 'invalid_request_error',
        },
      });
    }
  });

  app.post('/v1/update-agent-status', async (req, res) => {
    try {
      const { agentId, status, additionalData } = req.body;

      if (!agentId || !status) {
        return res.status(400).json({
          error: {
            message: 'Agent ID and status are required',
            type: 'invalid_request_error',
          },
        });
      }

      if (stabilityEngine) {
        stabilityEngine.updateAgentStatus(agentId, status, additionalData);
      }

      res.json({
        success: true,
        agentId: agentId,
        status: status,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/update-agent-status:', error);
      res.status(500).json({
        error: {
          message: 'Failed to update agent status',
          type: 'invalid_request_error',
        },
      });
    }
  });

  app.post('/v1/isolate-unstable-agents', async (req, res) => {
    try {
      if (!stabilityEngine) {
        return res.status(503).json({
          error: {
            message: 'Stability engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const isolationReport = stabilityEngine.isolateUnstableAgents();

      res.json({
        isolationReport: isolationReport,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/isolate-unstable-agents:', error);
      res.status(500).json({
        error: {
          message: 'Failed to isolate unstable agents',
          type: 'invalid_request_error',
        },
      });
    }
  });

  app.get('/v1/detect-conflicts', async (req, res) => {
    try {
      if (!stabilityEngine) {
        return res.status(503).json({
          error: {
            message: 'Stability engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const conflictReport = stabilityEngine.detectConflicts();

      res.json({
        conflictReport: conflictReport,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/detect-conflicts:', error);
      res.status(500).json({
        error: {
          message: 'Failed to detect conflicts',
          type: 'invalid_request_error',
        },
      });
    }
  });

  app.get('/v1/balance-load', async (req, res) => {
    try {
      if (!stabilityEngine) {
        return res.status(503).json({
          error: {
            message: 'Stability engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const loadDistribution = stabilityEngine.balanceLoad();

      res.json({
        loadDistribution: loadDistribution,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/balance-load:', error);
      res.status(500).json({
        error: {
          message: 'Failed to balance load',
          type: 'invalid_request_error',
        },
      });
    }
  });

  app.post('/v1/activate-protection-measures', async (req, res) => {
    try {
      if (!stabilityEngine) {
        return res.status(503).json({
          error: {
            message: 'Stability engine not initialized',
            type: 'service_unavailable',
          },
        });
      }

      const protectionResult = stabilityEngine.activateProtectionMeasures();

      res.json({
        protectionResult: protectionResult,
        timestamp: new Date().toISOString(),
      });
    } catch (error) {
      console.error('❌ Error in /v1/activate-protection-measures:', error);
      res.status(500).json({
        error: {
          message: 'Failed to activate protection measures',
          type: 'invalid_request_error',
        },
      });
    }
  });
}